the function return 32bits,  a left over since the swap to 64bits.  Making UL AES not being properly identified

+++ client/src/cmdhfmfu.c
@@ -1723,7 +1723,7 @@ out:
     return res;
 }
 
-uint32_t GetHF14AMfU_Type(void) {
+uint64_t GetHF14AMfU_Type(void) {
 
     uint64_t tagtype = MFU_TT_UNKNOWN;
     iso14a_card_select_t card;
